RateApiをhelperに移動 #41
